Feat: #2 Add basic scaffold
scaffold 및 topBar, bottomBar component 추가
